% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility_model_frame.R
\name{area_populations}
\alias{area_populations}
\title{Aggregate district level populations}
\usage{
area_populations(population, areas_wide, project, naomi_level)
}
\arguments{
\item{population}{Age/sex/space/time stratified population}

\item{areas_wide}{Area hierarchy in wide format}
}
\description{
Aggregate district level populations to all levels of area hierarchy
}
