% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_survey_fertility.R
\name{assign_cluster_area}
\alias{assign_cluster_area}
\title{Aggregate district level cluster dataframe}
\usage{
assign_cluster_area(clusters, areas_wide, area_level)
}
\arguments{
\item{clusters}{Dataframe of geolocated survey clusters}

\item{areas_wide}{Area hierarchy in wide format}

\item{area_level}{Integer referring to administrative level, where 0 is national, 1 is provincial etc.}
}
\value{
Dataframe list of clusters by administrative area, one list item per survey
}
\description{
Aggregate a dataframe of survey clusters geolocated to an administrative to any higher administrative level
}
\seealso{
\code{\link{get_areas}}
}
