% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_survey_fertility.R
\name{calculate_dhs_fertility}
\alias{calculate_dhs_fertility}
\title{Calculate fertility rates from DHS, MIS, AIS data}
\usage{
calculate_dhs_fertility(iso3, surveys, clusters, areas_wide)
}
\description{
Calculate fertility rates from DHS, MIS, AIS data
}
