% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_survey_fertility.R
\name{clusters_to_surveys}
\alias{clusters_to_surveys}
\title{Join individual recode survey datasets by cluster to area IDs}
\usage{
clusters_to_surveys(surveys, cluster_areas, single_tips = TRUE)
}
\arguments{
\item{surveys}{Survey dataframe as created by \code{rdhs::dhs_surveys()}}

\item{cluster_areas}{Dataframe of clusters assigned to areas. Output of \code{\link{cluster_areas}}.}

\item{single_tips}{desc this}
}
\value{
Nested list: \code{df$ir} and \code{df$tips_surv}. \code{df$ir} has one list item per survey/area combination and \code{df$tips_surv} has the corroborating TIPS value.
}
\description{
Join individual recode survey datasets by cluster to area IDs
}
\seealso{
\code{\link{cluster_areas}}
}
