% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics_survey_extract.R
\name{create_surveys_mics}
\alias{create_surveys_mics}
\title{Download MICS surveys}
\usage{
create_surveys_mics(iso3, mics_indicators)
}
\arguments{
\item{iso3}{ISO3 code}
}
\description{
Download MICS surveys
}
