% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics_survey_extract.R
\name{filter_mics}
\alias{filter_mics}
\title{Filter MICS datasets}
\usage{
filter_mics(dat, mics_indicators, survey_id_i)
}
\description{
Filter MICS household, women, and birth history datasets to key variables, and rename to ensure consistent column names between surveys
}
