% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{get_areas}
\alias{get_areas}
\title{Make area dataframes}
\usage{
get_areas(iso3_current, naomi_data_path, full = FALSE, wide = TRUE)
}
\arguments{
\item{iso3_current}{A string of one or more ISO-3 codes}

\item{naomi_data_path}{A path to directory containing input data files}

\item{full}{Boolean to return the full area heirarchy}

\item{wide}{Boolean to return the area hierarchy in wide format}
}
\description{
Create area hierarchy dataframes in the Naomi package format. The function will always return the area hierarchy in long format, with arguments additionally to return full and wide hierarchies
}
