% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Get populations}
\usage{
get_boundaries(iso3_current, naomi_data_path)
}
\arguments{
\item{iso3_current}{A string of one or more ISO-3 codes}

\item{naomi_data_path}{A path to directory containing input data files}
}
\description{
Returns a dataframe of area boundaries
}
