% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{get_input_files}
\alias{get_input_files}
\title{Extract area, population, boundary files from target directory}
\usage{
get_input_files(iso3_current, naomi_data_path)
}
\arguments{
\item{iso3_current}{A string of one or more ISO-3 codes}

\item{naomi_data_path}{A path to directory containing input data files}
}
\value{
A named list of file paths
}
\description{
Extract area, population, boundary files from target directory
}
