% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_survey_fertility.R
\name{ir_by_area}
\alias{ir_by_area}
\title{Join individual recode survey datasets by area}
\usage{
ir_by_area(ir, area_list, n, total)
}
\arguments{
\item{ir}{Individual recode survey dataset}

\item{area_list}{List of areas}
}
\description{
Join individual recode survey datasets by area
}
