% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics_survey_extract.R
\name{join_survey_areas}
\alias{join_survey_areas}
\title{Join household datasets with area hierarchy}
\usage{
join_survey_areas(fertility_mics_data, areas, warn = FALSE)
}
\description{
Join household datasets with area hierarchy
}
