% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics_survey_extract.R
\name{make_asfr_inputs}
\alias{make_asfr_inputs}
\title{Transform survey datasets into inputs for calc_asfr}
\usage{
make_asfr_inputs(mics_survey_areas, mics_survey_data)
}
\description{
Transform survey datasets into inputs for calc_asfr
}
