% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility_model_frame.R
\name{make_model_frames_batch}
\alias{make_model_frames_batch}
\title{Make model frames for batch running}
\usage{
make_model_frames_batch(lvl_map, population, asfr, areas_list, project = 2020)
}
\arguments{
\item{population}{Age/sex/space stratified population}

\item{asfr}{ASFRs by district and time}

\item{areas_list}{List of area files}

\item{project}{Model will by default produce estimates up to year of last survey. Integer projection year if desired, else FALSE}

\item{lvl}{Dataframe of district and province levels}
}
\description{
Create model frames and aggregation matrices for TMB model.
}
