% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility_model_frame.R
\name{make_model_frames_dev}
\alias{make_model_frames_dev}
\title{Make model frames}
\usage{
make_model_frames_dev(
  iso3,
  population,
  asfr,
  areas,
  naomi_level,
  project = 2020
)
}
\arguments{
\item{population}{Age/sex/space stratified population}

\item{asfr}{ASFRs by district and time}

\item{areas}{Area hierarchy}

\item{naomi_level}{Area level to produce estimates at}

\item{project}{Model will by default produce estimates up to year of last survey. Integer projection year if desired, else FALSE}
}
\description{
Create model frames and aggregation matrices for TMB model.
}
