% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility_model_frame.R
\name{make_rw_structure_matrix}
\alias{make_rw_structure_matrix}
\title{Make random walk structure matrices}
\usage{
make_rw_structure_matrix(x, order, adjust_diagonal = TRUE)
}
\arguments{
\item{x}{Matrix size}

\item{order}{Random walk order}

\item{adjust_diagonal}{Add 1E-6 to the matrix diagonal to make the matrix proper. Default = TRUE}
}
\description{
Make random walk structure matrices
}
