% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics_survey_extract.R
\name{transform_mics}
\alias{transform_mics}
\title{Transform MICS dataframes}
\usage{
transform_mics(mics_survey_data, mics_indicators)
}
\description{
Convert lists by survey to lists by dataset type
}
