% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_survey_fertility.R
\name{calculate_mics_fertility}
\alias{calculate_mics_fertility}
\title{Calculate fertility rates from MICS data}
\usage{
calculate_mics_fertility(iso3, mics_wm, mics_births_to_women)
}
\description{
Calculate fertility rates from MICS data
}
